Add some public bits
