+++ src/libmeasurement_kit/ooni/bouncer.cpp
@@ -8,14 +8,14 @@ namespace mk {
 namespace ooni {
 namespace bouncer {
 
-ErrorOr<Var<BouncerReply>> BouncerReply::create(std::string data, Var<Logger> logger) {
+ErrorOr<Var<BouncerReply>> BouncerReply::create(std::string data,
+                                                Var<Logger> logger) {
     Var<BouncerReply> reply(new BouncerReply);
     try {
         reply->response = nlohmann::json::parse(data);
         if (reply->response.find("error") != reply->response.end()) {
             if (reply->response["error"] == "collector-not-found") {
-                logger->warn(
-                        "collector not found for the requested test");
+                logger->warn("collector not found for the requested test");
                 return BouncerCollectorNotFoundError();
             }
             if (reply->response["error"] == "invalid-request") {
@@ -51,7 +51,7 @@ ErrorOr<std::string> BouncerReply::get_collector() {
     try {
         return get_base()["collector"];
     } catch (std::domain_error &) {
-        /* FALLTHROUGH */ ;
+        /* FALLTHROUGH */;
     }
     return BouncerValueNotFoundError();
 }
@@ -65,7 +65,7 @@ ErrorOr<std::string> BouncerReply::get_collector_alternate(std::string type) {
             }
         }
     } catch (std::domain_error &) {
-        /* FALLTHROUGH */ ;
+        /* FALLTHROUGH */;
     }
     return BouncerValueNotFoundError();
 }
@@ -74,7 +74,7 @@ ErrorOr<std::string> BouncerReply::get_name() {
     try {
         return get_base()["name"];
     } catch (std::domain_error &) {
-        /* FALLTHROUGH */ ;
+        /* FALLTHROUGH */;
     }
     return BouncerValueNotFoundError();
 }
@@ -83,12 +83,13 @@ ErrorOr<std::string> BouncerReply::get_test_helper(std::string name) {
     try {
         return get_base()["test-helpers"][name];
     } catch (std::domain_error &) {
-        /* FALLTHROUGH */ ;
+        /* FALLTHROUGH */;
     }
     return BouncerValueNotFoundError();
 }
 
-ErrorOr<std::string> BouncerReply::get_test_helper_alternate(std::string name, std::string type) {
+ErrorOr<std::string> BouncerReply::get_test_helper_alternate(std::string name,
+                                                             std::string type) {
     try {
         auto collectors = get_base()["test-helpers-alternate"][name];
         for (auto collector : collectors) {
@@ -97,19 +98,26 @@ ErrorOr<std::string> BouncerReply::get_test_helper_alternate(std::string name, s
             }
         }
     } catch (std::domain_error &) {
-        /* FALLTHROUGH */ ;
+        /* FALLTHROUGH */;
     }
     return BouncerValueNotFoundError();
 }
-
 
 ErrorOr<std::string> BouncerReply::get_version() {
     try {
         return get_base()["version"];
     } catch (std::domain_error &) {
-        /* FALLTHROUGH */ ;
+        /* FALLTHROUGH */;
     }
     return BouncerValueNotFoundError();
+}
+
+void post_net_tests(std::string base_bouncer_url, std::string test_name,
+                    std::string test_version, std::list<std::string> helpers,
+                    Callback<Error, Var<BouncerReply>> cb, Settings settings,
+                    Var<Reactor> reactor, Var<Logger> logger) {
+    post_net_tests_impl(base_bouncer_url, test_name, test_version, helpers, cb,
+                        settings, reactor, logger);
 }
 
 } // namespace mk
